Make Codec parse to tree, then interpret the graph
This is to help with the symmetry between encoding and decoding. The
basic process now looks like this:

    penman ----[parse]----> tree --[interpret]--> graph
    graph  --[configure]--> tree ----[format]---> penman

While the two-step parsing (three if you include lexing) may slow
things down a bit, there are some operations that may make sense to do
at the tree level, such as rearranging edges without restructuring, so
it may even be faster for some workflows. This claim is testable, so
some workflows should be benchmarked at some point to determine any
speed gains or losses.