Make Graph.epidata a member again
This has gone back and forth, but now Graph.epidata is a proper
member, not a property nor a function. Before it was a function (and
before that a property) because the epidata needed to be extracted
from the Graph's data, but now it is stored as a separate dictionary,
so there is no need to hide it behind the other accessors. Also, this
makes it easy to modify the epidata, which is desired for, e.g.,
reconfiguring layouts.