Fix type check errors and remove old code
This removes the layout ordering functions as they no longer work with
the current setup. They may be re-added in the future in a more
compatible way. This also removes the defunct layout.inspect()
function and changes the argument order on layout.configure().

Type checking is also fixed up in several places. Now the whole
codebase passes a typecheck with mypy.