Update Model class; require : in roles
This presents a possibly complete (for now) Model class with tests. As
part of this change, roles are now expected to have the initial
colon (:) character, which steps away from previous behavior in favor
of conformance with how many others talk about roles. There is still
some accounting of :-less roles: Model.canonicalize_role() will insert
the colon if it wasn't there, PENMANCodec.format() will check if a
role is the node-label role regardless of the presence of : and it
will also ensure a : exists on formatted roles. There may be another
place or two that behaves similarly, but in general the code does not
modify what it is given (a manually constructed Graph without : in its
roles will stay that way until serialization).