Add Tree class; parse comments in codec
This renames the old Tree type into the Node type so that Tree is the
name of a new class. It mainly just contains a node, but it also
serves as a place to contain metadata and define some tree-specific
methods.

The is_atomic() function now has a new home at penman.tree instead of
penman.layout.