Add graph.NODETYPE_ROLE; make roles always have :
Graphs can now determine that a triple is a node label because they
always use the penman.graph.NODETYPE_ROLE role instead of the
model-defined one. The model-defined one is now only for printing
triple conjunctions and some Model methods.

This commit also ensures that all roles that a Graph is instantiated
with have a leading : but manually adding colon-less roles by
appending to Graph.triples is possible.