Remove Graph.copy() and Graph.clear()
For copying, we can probably just rely on Python's copy.deepcopy().

For clearing, there was only one place that used it, and only to clear
epidata, so I replaced that call with the 3 lines it takes to do the
same thing. It can be added again if there's a need, but for now it's
good to have fewer functions to support.