Add a logger, --verbose and --quiet options
The logger is named 'penman', so module-level loggers named, e.g.,
'penman.layout', will inherit the 'penman' logger's settings. This
makes it so using penman.main (e.g., via the CLI) can set options that
don't get set when using it as a library.

By default, the loggers are set to the ERROR level, then -v, -vv, and
-vvv set it to WARNING, INFO, and DEBUG, respectively. The -q/--quiet
option turns off all output, including stdout, which is sometimes
useful for commands that test things, but may or may not be useful for
Penman. This setup may be re-evaluated later.