Model.reify() no longer inverts incoming triple
The idea was to indicate the in/out nature of reified edges without
creating epigraphical markers, but this turned out to be more trouble
than it was worth, and it wasn't even accurate if the original triple
is to be otherwise laid out as inverted. Now it just returns "pure"
triples and the caller (e.g., `transform.reify_edges()`) manages the
epigraphical markers.