Fix setup.py to avoid error when using tox
The docs/requirements.txt file is not copied into the files that tox
uses to test the code, but this file is read by setup.py, which tox
uses to install the module. This caused an error. One fix is to
include the file in a MANIFEST.in file, but I don't want to (a)
include it in installs and (b) have a MANIFEST.in file. This fix
avoids the issue by only reading the file if it is found.