Add layout.appears_inverted() function
Triples in a graph do not have a notion of inversion, but the
information about whether a triple was serialized as inverted, or will
possibly invert on serialization, exists in the epigraphical
data. This new function uses that information to give a hint as to the
inversion-status of a triple. It uses the same logic as
layout.configure(), so it should be reliable unless the graph was
modified.

Resolves #37