Process surface alignments in layout, not codec
Performance-wise, there is not a big difference (only improves parsing
speed about 3%). Also this creates a difference between the formal
grammar and the implemented one, and puts some parsing work outside of
lexing and parsing. The benefit is that Trees are much simpler and the
changes in `penman.layout.interpret()` and `penman.layout.configure()`
are easy to understand. In addition, both `penman.codec` and
`penman.tree` can be completely ignorant of the epigraph, which seems
right as it's something on top of the graph, not the tree.

Resolves #45