Add transform.dereify_edges()
And the --dereify-edges command-line option

This implementation tries to manage epigraphical data so that the
dereified information appears in the same spot as the incoming edge to
the formerly reified node, but there are some remaining
issues. Notably, if two adjacent "branches" get dereified, the final
POP from the previous branch makes the second one's triples appear at
earlier reentrancies, if any. What is needed is a way to remove a POP
from the end of a branch that gets reified, but guessing what is the
end of a branch is difficult when working with graph triples.