Partially revert change about alignment parsing
There was a performance gain with not specifically parsing alignments,
but it was problematic because the naive method for splitting off
those alignments did not consider ~ characters in strings. This commit
now has the lexer parse them again, but the codec will reattach them
to the role or atom target they went with. This commit does not return
to the tree edges with epidata, though.

Resolves #50