Add penman.constant module
This module recovers the ability to get constant values cast to their
appropriate datatypes, which are determined by the form of the
constant string. These functions are not called in normal operation
but are meant to be used by a user of the API who wants to work with
the values or who is constructing a graph and wants to ensure strings
are quoted appropriately.

Resolves #69